Move `CompletableCancelToken` out of tests into mtags.
This class is useful outside of tests.
